requirements.txt
requests
confluent-kafka
opensearch-py
autopep8
python-dotenv